refactor: User, BusinessProfile 리팩토링 및 메서드 검사 로직 추가(#147)
