+++ backend/src/main/java/com/minibeit/DataLoader.java
@@ -26,7 +26,7 @@ public class DataLoader implements CommandLineRunner {
     public void run(String... args) {
         if (userRepository.findAll().isEmpty()) {
             List<User> users = new ArrayList<>();
-            for (int i = 1; i < 21; i++) {
+            for (int i = 1; i < 1000; i++) {
                 User user = User.builder().oauthId(String.valueOf(i)).school(schoolRepository.findById(1L).get()).name("테스터" + i).job("테스트하는사람").gender(Gender.MALE).provider(SignupProvider.MINIBEIT).nickname("테스터" + i).phoneNum("010-1234-1234").role(Role.USER).birth(LocalDate.of(2000, 12, 12)).signupCheck(true).build();
                 users.add(user);
             }
