+++ backend/src/main/java/com/minibeit/common/component/LogAspect.java
@@ -1,0 +1,21 @@
+package com.minibeit.common.component;
+
+import lombok.extern.slf4j.Slf4j;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.springframework.stereotype.Component;
+
+@Component
+@Aspect
+@Slf4j
+public class LogAspect {
+    @Around("@annotation(LogExecutionTime)")
+    public Object logExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
+        long startTime = System.currentTimeMillis();
+        Object proceed = joinPoint.proceed();
+        long endTime = System.currentTimeMillis();
+        log.info("time : " + (double)(endTime - startTime) / 1000);
+        return proceed;
+    }
+}
