+++ backend/src/main/java/com/minibeit/common/component/LogExecutionTime.java
@@ -1,0 +1,11 @@
+package com.minibeit.common.component;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LogExecutionTime {
+}
