+++ backend/src/rest/AuthController.http
@@ -2,7 +2,7 @@ POST {{apiUrl}}/user/login
 Content-Type: application/json;charset=UTF-8
 
 {
-  "id": "2"
+  "id": "1"
 }
 
 > {%
@@ -18,11 +18,11 @@ Content-Type: multipart/form-data; boundary=boundary
 --boundary
 Content-Disposition: form-data; name="name";
 
-테스터1
+테스트0
 --boundary
 Content-Disposition: form-data; name="nickname";
 
-테스터022
+2
 --boundary
 Content-Disposition: form-data; name="gender";
