+++ backend/src/rest/BusinessProfileController.http
@@ -25,7 +25,7 @@ GET {{apiUrl}}/business/profile/mine/list
 Authorization: {{authorizationToken}}
 
 ###
-GET {{apiUrl}}/business/profile/1
+GET {{apiUrl}}/business/profile/2
 Authorization: {{authorizationToken}}
 
 ###
