+++ backend/src/test/java/com/minibeit/user/web/AuthControllerTest.java
@@ -42,52 +42,6 @@ class AuthControllerTest extends MvcTest {
     private RefreshTokenService refreshTokenService;
 
     @Test
-    @DisplayName("회원가입 문서화")
-    public void signup() throws Exception {
-        InputStream is = new ClassPathResource("mock/images/enjoy.png").getInputStream();
-        MockMultipartFile avatar = new MockMultipartFile("avatar", "avatar.jpg", "image/jpg", is.readAllBytes());
-
-        UserResponse.CreateOrUpdate response = UserResponse.CreateOrUpdate.builder().id(1L).nickname("동그라미").schoolId(1L).build();
-
-        given(authService.signup(any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(
-                multipart("/api/user/signup")
-                        .file(avatar)
-                        .param("name", "실명")
-                        .param("nickname", "동그라미")
-                        .param("gender", "MALE")
-                        .param("phoneNum", "010-1234-5678")
-                        .param("job", "대학생")
-                        .param("birth","2000-11-11")
-                        .param("schoolId", "1")
-                        .contentType(MediaType.MULTIPART_FORM_DATA)
-                        .characterEncoding("UTF-8")
-        );
-
-        results.andExpect(status().isOk())
-                .andDo(document("auth-signup",
-                        requestParameters(
-                                parameterWithName("name").description("실명"),
-                                parameterWithName("nickname").description("닉네임"),
-                                parameterWithName("gender").description("성별(MALE or FEMALE)"),
-                                parameterWithName("phoneNum").description("전화번호"),
-                                parameterWithName("birth").description("생년월일 (2000-11-11)"),
-                                parameterWithName("job").description("직업"),
-                                parameterWithName("schoolId").description("관심있는 학교 식별자")
-                        ),
-                        requestParts(
-                                partWithName("avatar").description("사용자 프로필 이미지")
-                        ),
-                        responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("회원가입한 유저 식별자"),
-                                fieldWithPath("nickname").type(JsonFieldType.STRING).description("회원가입한 유저 닉네임"),
-                                fieldWithPath("schoolId").type(JsonFieldType.NUMBER).description("관심 학교 식별자")
-                        )
-                ));
-    }
-
-    @Test
     @DisplayName("refresh token 문서화")
     public void refreshToken() throws Exception {
         UserResponse.Login response = UserResponse.Login.build(1L, "테스터",
